+++ internal/procgroup/group_test.go
@@ -22,6 +22,7 @@ const (
 	procgroupReadyEnv = "SPARKWING_PROCGROUP_READY"
 	procgroupReadyFD  = "SPARKWING_PROCGROUP_READY_FD"
 	procgroupTermSeen = "SPARKWING_PROCGROUP_TERM_SEEN"
+	procgroupOwnedPID = "SPARKWING_PROCGROUP_OWNED_PID"
 )
 
 func TestGroupHelperProcess(t *testing.T) {
@@ -59,6 +60,13 @@ func TestGroupHelperProcess(t *testing.T) {
 	case "session-parked":
 		IgnoreTermination()
 		holdHelperProcess("")
+	case "owner":
+		ForwardTerminationToOwned()
+		if err := startOwnedSessionDescendant(os.Getenv(procgroupOwnedPID)); err != nil {
+			_, _ = fmt.Fprintln(os.Stderr, err)
+			os.Exit(2)
+		}
+		holdHelperProcess(os.Getenv(procgroupReadyEnv))
 	case "session-cooperative":
 		term := make(chan os.Signal, 1)
 		release := make(chan os.Signal, 1)